backend/requirements.txt
fastapi==0.115.8
uvicorn==0.27.1
sqlalchemy==2.0.37
pydantic==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.14.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==79.0.1